feat(component): add ConfirmIconButton and update file card delete button
address #150

ConfirmIconButton extends the IconButton with a ConfirmationDialog. If
the confirmationDescription and confirmationTitle exist, a dialog will
shown when the button onClick event is triggered.  In the
ConfirmationDialog, if the cancel button is clicked, it will close and
ignore the onClick event of the ConfirmIconButton.  If the confirm
button is clicked, the onClick event of the ConfirmIconButton will be
triggered.